requirements.txt
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
idna==3.4
pathlib==1.0.1
pydantic==1.10.7
Pygments==2.14.0
requests==2.28.2
soupsieve==2.4
typing_extensions==4.5.0
urllib3==1.26.15
azure-cognitiveservices-speech==1.30.0
